requirements.txt
sounddevice
soundfile
whisper
torch
pyttsx4
opencv-python
Pillow
firebase-admin
requests
openai
google-generativeai
stepic
pyaudio
SpeechRecognition
PyQt5
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git